Move type validation into a sub-module
